341: Move the FRFrequency classes in securebanking-openbanking-uk-common
Issue: https://github.com/SecureBankingAccessToolkit/SecureBankingAccessToolkit/issues/341
