+++ securebanking-openbanking-uk-common/src/main/java/com/forgerock/securebanking/openbanking/uk/common/api/meta/obie/OBGroupName.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.forgerock.securebanking.openbanking.uk.common.api.meta;
+package com.forgerock.securebanking.openbanking.uk.common.api.meta.obie;
 
 public enum OBGroupName {
