+++ securebanking-openbanking-uk-obie-datamodel/pom.xml
@@ -119,6 +119,12 @@
             <artifactId>assertj-core</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>com.forgerock.securebanking.uk</groupId>
+            <artifactId>securebanking-openbanking-uk-common</artifactId>
+            <version>1.1.3-SNAPSHOT</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
